ucid-app-clean/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
Werkzeug==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
anthropic==0.19.1
openai==1.12.0
